# src/admin/FocusTemplate.Admin.Api/generated/UserManagement/Groups/Endpoints/GroupsEndpoints.cs
// <auto-generated/>

#nullable enable

using System.Collections.Generic;
using System.Threading;
using System.Threading.Tasks;
using FocusTemplate.Admin.Shared;
using FocusTemplate.Admin.Shared.UserManagement;
using FocusTemplate.Primitives;
using Mediator;
using Microsoft.AspNetCore.Builder;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Http.HttpResults;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Routing;
using ValidationProblemDetails = FocusTemplate.Admin.Shared.ValidationProblemDetails;

namespace FocusTemplate.Admin.Api.Features.UserManagement.Groups;

/// <summary>Minimal API endpoints for Groups, generated from the TypeSpec contract.</summary>
public static partial class GroupsEndpoints
{
	/// <summary>Maps the Groups endpoints.</summary>
	/// <param name="app">The route builder to add the endpoints to.</param>
	/// <returns>The route group the endpoints were added to, for further group-wide configuration.</returns>
	public static RouteGroupBuilder MapGroupsEndpoints(this IEndpointRouteBuilder app)
	{
		ValidationProblems.EnsureRegistered<GroupsCreateCommand>(app.ServiceProvider);

		RouteGroupBuilder group = app.MapGroup("/groups").WithTags("Groups");
		ConfigureGroup(group);

		RouteHandlerBuilder listEndpoint = group.MapGet("", ListHandler).WithName("UserManagement.Groups_list").Produces<IReadOnlyList<GroupSummaryResponse>>(200).Produces(401).Produces(403).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ViewGroups));
		ConfigureList(listEndpoint);

		RouteHandlerBuilder getEndpoint = group.MapGet("/{id:guid}", GetHandler).WithName("UserManagement.Groups_get").Produces<GroupResponse>(200).ProducesProblem(404).Produces(401).Produces(403).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ViewGroups));
		ConfigureGet(getEndpoint);

		RouteHandlerBuilder createEndpoint = group.MapPost("", CreateHandler).WithName("UserManagement.Groups_create").Produces<GroupResponse>(200).Produces<ValidationProblemDetails>(400, "application/problem+json").ProducesProblem(409).Produces(401).Produces(403).AddEndpointFilter(ValidateCreate).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ManageGroups));
		ConfigureCreate(createEndpoint);

		RouteHandlerBuilder updateEndpoint = group.MapPut("/{id:guid}", UpdateHandler).WithName("UserManagement.Groups_update").Produces<GroupResponse>(200).Produces<ValidationProblemDetails>(400, "application/problem+json").ProducesProblem(404).ProducesProblem(409).Produces(401).Produces(403).AddEndpointFilter(ValidateUpdate).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ManageGroups));
		ConfigureUpdate(updateEndpoint);

		RouteHandlerBuilder deleteEndpoint = group.MapDelete("/{id:guid}", DeleteHandler).WithName("UserManagement.Groups_delete").Produces(204).ProducesProblem(409).Produces(401).Produces(403).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ManageGroups));
		ConfigureDelete(deleteEndpoint);

		RouteHandlerBuilder setPermissionsEndpoint = group.MapPut("/{id:guid}/permissions", SetPermissionsHandler).WithName("UserManagement.Groups_setPermissions").Produces(204).ProducesProblem(404).ProducesProblem(409).Produces(401).Produces(403).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ManageGroups));
		ConfigureSetPermissions(setPermissionsEndpoint);

		RouteHandlerBuilder addMemberEndpoint = group.MapPut("/{id:guid}/members/{userId:guid}", AddMemberHandler).WithName("UserManagement.Groups_addMember").Produces(204).ProducesProblem(404).Produces(401).Produces(403).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ManageGroups));
		ConfigureAddMember(addMemberEndpoint);

		RouteHandlerBuilder removeMemberEndpoint = group.MapDelete("/{id:guid}/members/{userId:guid}", RemoveMemberHandler).WithName("UserManagement.Groups_removeMember").Produces(204).ProducesProblem(409).Produces(401).Produces(403).RequireAuthorization().WithMetadata(new RequiresPermissionAttribute(global::FocusTemplate.Primitives.UserManagementPermissions.ManageGroups));
		ConfigureRemoveMember(removeMemberEndpoint);

		return group;
	}

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static ValueTask<IReadOnlyList<GroupSummaryResponse>> ListHandler(IMediator mediator, CancellationToken cancellationToken = default) =>
		mediator.Send(new GroupsListQuery(), cancellationToken);

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="id">id.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static async Task<Results<Ok<GroupResponse>, ProblemHttpResult>> GetHandler(IMediator mediator, GroupId id, CancellationToken cancellationToken = default)
	{
		GroupsGetResult result = await mediator.Send(new GroupsGetQuery(id), cancellationToken);
		return result.ToHttpResult();
	}

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="body">body.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static async Task<Results<Ok<GroupResponse>, ProblemHttpResult, ValidationProblemResult>> CreateHandler(IMediator mediator, [FromBody] GroupRequest body, CancellationToken cancellationToken = default)
	{
		GroupsCreateResult result = await mediator.Send(new GroupsCreateCommand(body), cancellationToken);
		return result.ToHttpResult();
	}

	/// <summary>The validation filter of the create endpoint.</summary>
	private static ValueTask<object?> ValidateCreate(EndpointFilterInvocationContext context, EndpointFilterDelegate next) =>
		ValidationProblems.ValidateAsync(context, next, new GroupsCreateCommand(context.GetArgument<GroupRequest>(1)));

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="id">id.</param>
	/// <param name="body">body.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static async Task<Results<Ok<GroupResponse>, ProblemHttpResult, ValidationProblemResult>> UpdateHandler(IMediator mediator, GroupId id, [FromBody] GroupRequest body, CancellationToken cancellationToken = default)
	{
		GroupsUpdateResult result = await mediator.Send(new GroupsUpdateCommand(id, body), cancellationToken);
		return result.ToHttpResult();
	}

	/// <summary>The validation filter of the update endpoint.</summary>
	private static ValueTask<object?> ValidateUpdate(EndpointFilterInvocationContext context, EndpointFilterDelegate next) =>
		ValidationProblems.ValidateAsync(context, next, new GroupsUpdateCommand(context.GetArgument<GroupId>(1), context.GetArgument<GroupRequest>(2)));

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="id">id.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static async Task<Results<NoContent, ProblemHttpResult>> DeleteHandler(IMediator mediator, GroupId id, CancellationToken cancellationToken = default)
	{
		GroupsDeleteResult result = await mediator.Send(new GroupsDeleteCommand(id), cancellationToken);
		return result.ToHttpResult();
	}

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="id">id.</param>
	/// <param name="permissions">permissions.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static async Task<Results<NoContent, ProblemHttpResult>> SetPermissionsHandler(IMediator mediator, GroupId id, [FromBody] IReadOnlyList<Permission> permissions, CancellationToken cancellationToken = default)
	{
		GroupsSetPermissionsResult result = await mediator.Send(new GroupsSetPermissionsCommand(id, permissions), cancellationToken);
		return result.ToHttpResult();
	}

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="id">id.</param>
	/// <param name="userId">userId.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static async Task<Results<NoContent, ProblemHttpResult>> AddMemberHandler(IMediator mediator, GroupId id, UserId userId, CancellationToken cancellationToken = default)
	{
		GroupsAddMemberResult result = await mediator.Send(new GroupsAddMemberCommand(id, userId), cancellationToken);
		return result.ToHttpResult();
	}

	/// <param name="mediator">The Mediator instance.</param>
	/// <param name="id">id.</param>
	/// <param name="userId">userId.</param>
	/// <param name="cancellationToken">Cancellation token.</param>
	private static async Task<Results<NoContent, ProblemHttpResult>> RemoveMemberHandler(IMediator mediator, GroupId id, UserId userId, CancellationToken cancellationToken = default)
	{
		GroupsRemoveMemberResult result = await mediator.Send(new GroupsRemoveMemberCommand(id, userId), cancellationToken);
		return result.ToHttpResult();
	}

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to every Groups endpoint at once; the Configure{Op} hooks refine single endpoints.</summary>
	/// <param name="group">The route group of the Groups endpoints.</param>
	static partial void ConfigureGroup(RouteGroupBuilder group);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the list endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureList(RouteHandlerBuilder builder);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the get endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureGet(RouteHandlerBuilder builder);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the create endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureCreate(RouteHandlerBuilder builder);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the update endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureUpdate(RouteHandlerBuilder builder);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the delete endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureDelete(RouteHandlerBuilder builder);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the setPermissions endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureSetPermissions(RouteHandlerBuilder builder);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the addMember endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureAddMember(RouteHandlerBuilder builder);

	/// <summary>Implement in a hand-written partial to attach policies (authorization, rate limiting, caching) to the removeMember endpoint.</summary>
	/// <param name="builder">The route handler builder of the endpoint.</param>
	static partial void ConfigureRemoveMember(RouteHandlerBuilder builder);
}