# src/admin/FocusTemplate.Admin.Api/openapi.yaml
openapi: 3.2.0
info:
  title: FocusTemplate Admin API
  version: 0.1.0
  description: Internal API of the Admin vertical. Reached only through the BFF, never by a browser directly.
...
paths:
...
  /groups:
    get:
      operationId: Groups_list
      description: Every group, by name.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserManagement.GroupSummaryResponse'
      tags:
        - Groups
    post:
      operationId: Groups_create
      description: Creates a group without members or permissions.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserManagement.GroupResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/SpatialFocus.Http.ValidationProblemDetails'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserManagement.GroupRequest'
  /groups/{id}:
    get:
      operationId: Groups_get
      description: A group with its members and permissions.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/GroupId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserManagement.GroupResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Groups
    put:
      operationId: Groups_update
      description: Renames a group or changes its description, a managed group too.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/GroupId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserManagement.GroupResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/SpatialFocus.Http.ValidationProblemDetails'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserManagement.GroupRequest'
    delete:
      operationId: Groups_delete
      description: Deletes a group, and its members lose what it granted. Not a managed group. A missing group answers 204 too, so a retry stays idempotent.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/GroupId'
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Groups
  /groups/{id}/members/{userId}:
    put:
      operationId: Groups_addMember
      description: Adds a user to the group. A member already stays one.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/GroupId'
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserId'
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '404':
          description: The server cannot find the requested resource.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Groups
    delete:
      operationId: Groups_removeMember
      description: Removes a user from the group. One who is no member answers 204 too. Not the last active member of the Administrators.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/GroupId'
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserId'
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Groups
  /groups/{id}/permissions:
    put:
      operationId: Groups_setPermissions
      description: Sets every permission the group grants, replacing the ones before. Not for a managed group.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/GroupId'
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '404':
          description: The server cannot find the requested resource.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Permission'
...
security:
  - BearerAuth: []
components:
  schemas:
...
    GroupId:
      type: string
      format: uuid
      description: Identifier of a group of users, the unit permissions are granted to.
    Permission:
      type: string
      description: A permission, named after its feature slice and action, such as UserManagement.ManageGroups. Admins grant permissions to groups.
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        detail:
          type: string
          description: Explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference identifying this occurrence.
      unevaluatedProperties: {}
      description: |-
        Problem details for an HTTP API response, as defined by RFC 9457
        (https://www.rfc-editor.org/rfc/rfc9457). Extension members are allowed and carry information
        specific to the problem type, such as a trace id or the fields of a validation failure.
    SpatialFocus.Http.ValidationProblemDetails:
      type: object
      required:
        - errors
        - violations
      properties:
        type:
          type: string
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        detail:
          type: string
          description: Explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference identifying this occurrence.
        errors:
          type: object
          unevaluatedProperties:
            type: array
            items:
              type: string
          description: Error messages by wire path (name, address.street, lines[2].quantity), errors only; the shape of ASP.NET Core's HttpValidationProblemDetails.
        violations:
          type: array
          items:
            $ref: '#/components/schemas/SpatialFocus.Http.Violation'
          description: Every failed rule, warnings and infos included.
      unevaluatedProperties: {}
      description: |-
        Problem details (RFC 9457) of a request that failed validation: the errors by the wire path of the member or parameter
        they concern, and every failed rule with its code, severity and the constraint's values.
    SpatialFocus.Http.Violation:
      type: object
      required:
        - key
        - code
        - severity
        - message
      properties:
        key:
          type: string
          description: Wire path of the member or parameter, as in the errors of the problem.
        code:
          type: string
          description: "What failed: for a rule of the contract the constraint's name (required, minLength, maxLength, pattern, minValue, maxValue, minValueExclusive, maxValueExclusive, minItems, maxItems), otherwise the code the rule sets."
        severity:
          $ref: '#/components/schemas/SpatialFocus.Http.ViolationSeverity'
          description: Only an error blocks the request.
        message:
          type: string
          description: The message, in the language of the server.
        args:
          type: object
          unevaluatedProperties: {}
          description: The constraint's values by name, maxLength for example, for a client that phrases its own message.
      description: One failed rule of a validation.
    SpatialFocus.Http.ViolationSeverity:
      type: string
      enum:
        - error
        - warning
        - info
      description: 'How much a failed rule weighs: only an error blocks the request.'
    UserId:
      type: string
      format: uuid
      description: Identifier of a user, the owner of stations and the author of audited changes.
...
    UserManagement.GroupMemberResponse:
      type: object
      required:
        - id
        - displayName
        - isActive
      properties:
        id:
          $ref: '#/components/schemas/UserId'
          description: Identifier of the user.
        displayName:
          type: string
          description: Display name.
        isActive:
          type: boolean
          description: False once the user is deactivated.
      description: A member of a group.
...
    UserManagement.GroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name, unique.
        description:
          type: string
          maxLength: 500
          description: What the group is for.
      description: The name and description of a group.
    UserManagement.GroupResponse:
      type: object
      required:
        - id
        - name
        - isManaged
        - memberCount
        - members
        - permissions
      properties:
        id:
          $ref: '#/components/schemas/GroupId'
          description: Identifier.
        name:
          type: string
          description: Name, unique.
        description:
          type: string
          description: What the group is for.
        isManaged:
          type: boolean
          description: 'Managed by the application: its permissions come with each release, and it cannot be deleted.'
        memberCount:
          type: integer
          format: int32
          description: How many users are members.
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserManagement.GroupMemberResponse'
          description: The members, by display name.
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
          description: The permissions the group grants its members.
      description: A group with its members and permissions.
    UserManagement.GroupSummaryResponse:
      type: object
      required:
        - id
        - name
        - isManaged
        - memberCount
      properties:
        id:
          $ref: '#/components/schemas/GroupId'
          description: Identifier.
        name:
          type: string
          description: Name, unique.
        description:
          type: string
          description: What the group is for.
        isManaged:
          type: boolean
          description: 'Managed by the application: its permissions come with each release, and it cannot be deleted.'
        memberCount:
          type: integer
          format: int32
          description: How many users are members.
      description: A group in a list.
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer